Add ActionsHandler and move actions code to this new class
